Get tweets data for statuses favorited by one or more target users.

Returns up to 3,000 statuses favorited by each of one or more specific Twitter users.

get_favorites(user, n = 200, since_id = NULL, max_id = NULL,
  parse = TRUE, token = NULL)

Arguments

user
Vector of user names, user IDs, or a mixture of both.

n
Specifies the number of records to retrieve. Defaults to 200. 3000 is the max number of favorites returned per token. Due to suspended or deleted content, this function may return fewer tweets than the desired (n) number. Must be of length 1 or of length equal to the provided number of users.

since_id
Returns results with an status_id greater than (that is, more recent than) the specified status_id. There are limits to the number of tweets returned by the REST API. If the limit is hit, since_id is adjusted (by Twitter) to the oldest ID available.

max_id
Returns results with status_id less (older) than or equal to (if hit limit) the specified status_id.

parse
Logical, indicating whether to return parsed vector or nested list object. By default, parse = TRUE saves you the time [and frustrations] associated with disentangling the Twitter API return objects.

token
OAuth token. By default token = NULL fetches a non-exhausted token from an environment variable. Find instructions on how to create tokens and setup an environment variable in the tokens vignette (in r, send ?tokens to console).

Value

A tbl data frame of tweets data with users data attribute.

See also

https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/get-favorites-list

Other tweets: get_mentions, get_timeline, lists_statuses, lookup_statuses, search_tweets, tweets_data, tweets_with_users

Examples

# NOT RUN {
## get max number of statuses favorited by KFC
kfc <- get_favorites("KFC", n = 3000)
kfc

## get 400 statuses favorited by each of three users
favs <- get_favorites(c("Lesdoggg", "pattonoswalt", "meganamram"))
favs

# }