GET friends/ids

Returns user IDs of accounts followed by target user.

get_friends(users, retryonratelimit = FALSE, ...)

Arguments

users
Screen name or user ID of target user from which the user IDs of friends (accounts followed BY target user) will be retrieved.

retryonratelimit
If you'd like to retrieve 5,000 or fewer friends for more than 15 target users, then set retryonratelimit = TRUE and this function will use Sys.sleep until rate limits reset and the desired number of friend networks is retrieved. This defaults to FALSE. See details for more info regarding possible issues with timing misfires.

...
For other possible args see get_friends.default.

Value

A tibble data frame with two columns, "user" for name or ID of target user and "user_id" for follower IDs.

Details

When retryonratelimit = TRUE this function internally makes a rate limit API call to get information on (a) the number of requests remaining and (b) the amount of time until the rate limit resets. So, in theory, the sleep call should only be called once between waves of data collection. However, as a fail safe, if a system's time is calibrated such that it expires before the rate limit reset, or if, in another session, the user dips into the rate limit, then this function will wait (use Sys.sleep for a second time) until the next rate limit reset. Users should monitor and test this before making especially large calls as any systematic issues could create sizable inefficiencies.