Get your timeline

Returns a collection of the most recent Tweets and Retweets posted by the authenticating user and the users they follow. The home timeline is central to how most users interact with the Twitter service. The authenticating user is determined from the token.

get_my_timeline(n = 100, max_id = NULL, parse = TRUE, check = TRUE,
  token = NULL, ...)

Arguments

n
Number of tweets to return per timeline. Defaults to 100. Must be of length 1 or equal to length of user.

max_id
Character, status_id from which returned tweets should be older than.

parse
Logical, indicating whether to return parsed (data.frames) or nested list object. By default, parse = TRUE saves users from the time (and frustrations) associated with disentangling the Twitter API return objects.

check
Logical indicating whether to remove check available rate limit. Ensures the request does not exceed the maximum remaining number of calls. Defaults to TRUE.

token
OAuth token. By default token = NULL fetches a non-exhausted token from an environment variable. Find instructions on how to create tokens and setup an environment variable in the tokens vignette (in r, send ?tokens to console).

...
Further arguments passed on as parameters in API query.

Value

A tbl data frame of tweets data with users data attribute.

See also

https://developer.twitter.com/en/docs/tweets/timelines/api-reference/get-statuses-home_timeline

Other tweets: get_favorites, get_mentions, get_timeline, lists_statuses, lookup_statuses, search_tweets, tweets_data, tweets_with_users

Examples

# NOT RUN {
tweets_from_me_and_the_ppl_i_follow <- get_my_timeline(n = 3200)

# }