Get user IDs of users who retweeted a given status.

Returns user IDs of users who retweeted a given status. At the current time, this function is limited in returning a maximum of 100 users for a given status.

get_retweeters(status_id, n = 100, parse = TRUE, token = NULL)

Arguments

status_id
required The status ID of the desired status.

n
Specifies the number of records to retrieve. Best if intervals of 100.

parse
Logical indicating whether to convert the response object into an R list. Defaults to TRUE.

token
OAuth token. By default token = NULL fetches a non-exhausted token from an environment variable. Find instructions on how to create tokens and setup an environment variable in the tokens vignette (in r, send ?tokens to console).

Value

data

Details

At time of writing, pagination offers no additional data. See the post from Pipes here: https://twittercommunity.com/t/paging-is-not-possible-with-statuses-retweeters-ids-json/71298/8

See also

Other retweets: get_retweets