Get a timeline of tweets authored by members of a specified list.

Get a timeline of tweets authored by members of a specified list.

lists_statuses(list_id = NULL, slug = NULL, owner_user = NULL,
  since_id = NULL, max_id = NULL, n = 5000, include_rts = TRUE,
  parse = TRUE, token = NULL)

Arguments

list_id
required The numerical id of the list.

slug
required You can identify a list by its slug instead of its numerical id. If you decide to do so, note that you'll also have to specify the list owner using the owner_id or owner_screen_name parameters.

owner_user
optional The screen name or user ID of the user who owns the list being requested by a slug.

since_id
optional Returns results with an ID greater than (that is, more recent than) the specified ID. There are limits to the number of Tweets which can be accessed through the API. If the limit of Tweets has occurred since the since_id, the since_id will be forced to the oldest ID available.

max_id
optional Returns results with an ID less than (that is, older than) or equal to the specified ID.

n
optional Specifies the number of results to retrieve per "page."

include_rts
optional When set to either true, t or 1, the list timeline will contain native retweets (if they exist) in addition to the standard stream of tweets. The output format of retweeted tweets is identical to the representation you see in home_timeline.

parse
Logical indicating whether to convert the response object into an R list. Defaults to TRUE.

token
OAuth token. By default token = NULL fetches a non-exhausted token from an environment variable. Find instructions on how to create tokens and setup an environment variable in the tokens vignette (in r, send ?tokens to console).

Value

data

See also

Other lists: lists_members, lists_subscribers, lists_users

Other tweets: get_favorites, get_mentions, get_timeline, lookup_statuses, search_tweets, tweets_data, tweets_with_users