Get collections by user or status id.

Return data for specified collection (themed grouping of Twitter statuses). Response data varies significantly compared to most other users and tweets data objects returned in this package.

lookup_collections(id, n = 200, parse = TRUE, token = NULL, ...)

Arguments

id
required The identifier of the Collection to return results for e.g., "custom-539487832448843776"

n
Specifies the maximum number of results to include in the response. Specify count between 1 and 200.

parse
Logical indicating whether to convert response object into nested list. Defaults to true.

token
OAuth token. By default token = NULL fetches a non-exhausted token from an environment variable. Find instructions on how to create tokens and setup an environment variable in the tokens vignette (in r, send ?tokens to console).

...
Other arguments passed along to composed request query.

Value

Return object converted to nested list if parsed otherwise an HTTP response object is returned.

Examples

# NOT RUN {
## lookup a specific collection
cc <- lookup_collections("custom-539487832448843776")

## inspect data
str(cc)

# }