Get coordinates of specified location.

Convenience function for looking up latitude/longitude coordinate information for a given location. Returns data as a special "coords" object, which is specifically designed to interact smoothly with other relevant package functions.

lookup_coords(address, components = NULL, ...)

Arguments

address
Desired location typically in the form of place name, subregion, e.g., address = "lawrence, KS". Also accepts the name of countries, e.g., address = "usa", address = "brazil" or states, e.g., address = "missouri" or cities, e.g., address = "chicago". In most cases using only address should be sufficient.

components
Unit of analysis for address e.g., components = "country:US". Potential components include postal_code, country, administrative_area, locality, route.

...
Additional arguments passed as parameters in the HTTP request

Value

Object of class coords.

See also

Other geo: lat_lng

Examples

# NOT RUN {
## get coordinates associated with the following addresses/components
sf <- lookup_coords("san francisco, CA", "country:US")
usa <- lookup_coords("usa")
lnd <- lookup_coords("london")
bz <- lookup_coords("brazil")

## pass a returned coords object to search_tweets
bztw <- search_tweets(geocode = bz)

## or stream tweets
ustw <- stream_tweets(usa, timeout = 10)

# }