ply_friends

ply_friends

ply_friends(users, token = NULL, ...)

Arguments

users

Screen names and/or user ids of target user. Since friend networks are requested one at a time, users can be a mixture of user_ids and screen_names.

token

OAuth token. By default token = NULL fetches a non-exhausted token from an environment variable. Find instructions on how to create tokens and setup an environment variable in the tokens vignette (in r, send ?tokens to console).

…

Arguments passed on to get_friends.

Value

List of friend networks where nth element contains user ids of nth users.

Examples

## Not run: ------------------------------------
# get friend networks for up to 15 users per token
# users <- c("inside_R", "rtweet_package", "RLangTip", "Rbloggers", "rstudio")
# friend.networks <- ply_friends(users)
# friend.networks[[1]]
# str(friend.networks)
## ---------------------------------------------