stopwordslangs data

This data comes form a group of Twitter searches conducted on 2017-09-27. The data are commonly observed words associated with 10 different languages, including "ar", "en", "es", "fr", "in", "ja", "pt", ru", "tr", and "und". Variables include term (the stop words), lang, and z (standard deviation units of logged word occurence in stopwords sample).

stopwordslangs

Format

A tibble with three variables and 2,931 observations

Examples

stop_words("en")

#>   [1] "the"    "to"     "i"      "a"      "you"    "and"    "of"     "is"    
#>   [9] "in"     "for"    "my"     "it"     "this"   "on"     "that"   "me"    
#>  [17] "with"   "im"     "be"     "so"     "are"    "not"    "have"   "but"   
#>  [25] "your"   "at"     "just"   "its"    "like"   "all"    "dont"   "was"   
#>  [33] "if"     "can"    "get"    "they"   "what"   "amp"    "one"    "we"    
#>  [41] "about"  "no"     "out"    "he"     "do"     "now"    "from"   "love"  
#>  [49] "will"   "up"     "as"     "how"    "by"     "new"    "when"   "good"  
#>  [57] "know"   "people" "time"   "go"     "our"    "or"     "more"   "make"  
#>  [65] "need"   "an"     "see"    "think"  "why"    "has"    "who"    "day"   
#>  [73] "some"   "cant"   "u"      "would"  "lol"    "want"   "thats"  "them"  
#>  [81] "been"   "his"    "there"  "still"  "today"  "got"    "only"   "really"
#>  [89] "youre"  "us"     "were"   "her"    "she"    "too"    "back"   "here"  
#>  [97] "should" "right"  "well"   "thank" 
#>  [ reached getOption("max.print") -- omitted 2831 entries ]