Parsing data into tweets/users data tibbles

Parsing data into tweets/users data tibbles

tweets_with_users(x)

users_with_tweets(x)

Arguments

x
Unparsed data returned by rtweet API request.

Value

A tweets/users tibble (data frame) with users/tweets tibble attribute.

See also

Examples

if (FALSE) {
## search with parse = FALSE
rt <- search_tweets("rstats", n = 500, parse = FALSE)

## parse to tweets data tibble with users data attribute object
tweets_with_users(rt)

## search with parse = FALSE
usr <- search_users("rstats", n = 300, parse = FALSE)

## parse to users data tibble with users data attribute object
users_with_tweets(usr)

}