users_data

Returns users data frame from returned tweets data object.

users_data(tweets)

Arguments

tweets

Data frame of Twitter statuses (tweets) generated via get_timeline, search_tweets, or stream_tweets.

Value

Users data frame from tweets returned in a tweets data object.

See also

Other users: lookup_users, search_users

Examples

## Not run: ------------------------------------
# # search for 100 tweets containing the letter r
# r <- search_tweets("r")
# 
# # print tweets data (only first 10 rows are shown)
# r
# 
# # extract users data
# users_data(r)
## ---------------------------------------------