Get collections by user or status id.

Find collections (themed grouping of statuses) created by specific user or status id. Results include user, status, and collection features.

get_collections(user, status_id = NULL, n = 200, cursor = NULL,
  parse = TRUE, token = NULL)

Arguments

user
Screen name or user id of target user. Requests must provide a value for one of user or status_id.

status_id
Optional, the identifier of the tweet for which to return results. Requests must provide a value for one of user or status_id.

n
Maximum number of results to return. Defaults to 200.

cursor
Page identifier of results to retrieve. If parse = TRUE, the next cursor value for any given request--if available--is stored as an attribute, accessible via next_cursor

parse
Logical indicating whether to convert response object into nested list. Defaults to true.

token
OAuth token. By default token = NULL fetches a non-exhausted token from an environment variable. Find instructions on how to create tokens and setup an environment variable in the tokens vignette (in r, send ?tokens to console).

Value

Return object converted to nested list if parsed otherwise an HTTP response object is returned.

Examples

# NOT RUN {
## lookup a specific collection
cnnc <- get_collections("cnn")

## inspect data
str(cnnc)

## by status id
wwe <- get_collections(status_id = "925172982313570306")

## inspect data
str(wwe)

# }