Get the most recent retweets/retweeters

get_retweets() returns the 100 most recent retweets of a tweet; get_retweeters() returns the 100 most recent users who retweeted them.

Usage

get_retweets(status_id, n = 100, parse = TRUE, token = NULL, ...)

get_retweeters(status_id, n = 100, parse = TRUE, token = NULL)

Arguments

status_id

Tweet/status id.

n

Number of results to retrieve. Must be <= 100.

parse

If TRUE, the default, returns a tidy data frame. Use FALSE to return the "raw" list corresponding to the JSON returned from the Twitter API.

token

Use this to override authentication for a single API call. In many cases you are better off changing the default for all calls. See auth_as() for details.

...

Other arguments used as parameters in the query sent to Twitter's rest API, for example, trim_user = TRUE.

Value

Tweets data of the most recent retweets/retweeters of a given status.

References

https://developer.twitter.com/en/docs/twitter-api/v1/tweets/post-and-engage/api-reference/get-statuses-retweets-id

https://developer.twitter.com/en/docs/twitter-api/v1/tweets/post-and-engage/api-reference/get-statuses-retweeters-ids