GET statuses/retweets/:id

Returns a collection of the 100 most recent retweets of the Tweet specified by the id parameter. Twitter's API is currently limited to 100 or fewer retweeters.

get_retweets(status_id, n = 100, parse = TRUE, ..., token = NULL)

Arguments

status_id
required The numerical ID of the desired status.

n
optional Specifies the number of records to retrieve. Must be less than or equal to 100.

parse
Logical indicating whether to convert the response object into an R list. Defaults to TRUE.

...
Other arguments used as parameters in the query sent to Twitter's rest API, for example, trim_user = TRUE.

token
OAuth token. By default token = NULL fetches a non-exhausted token from an environment variable. Find instructions on how to create tokens and setup an environment variable in the tokens vignette (in r, send ?tokens to console).

Value

data