Get tweets data for given statuses (status IDs).

Returns data on up to 90,000 Twitter statuses. To return data on more than 90,000 statuses, users must iterate through status IDs whilst avoiding rate limits, which reset every 15 minutes.

lookup_statuses(statuses, parse = TRUE, token = NULL)

lookup_tweets(statuses, parse = TRUE, token = NULL)

Arguments

statuses
User id or screen name of target user.

parse
Logical, indicating whether or not to parse return object into data frame(s).

token
OAuth token. By default token = NULL fetches a non-exhausted token from an environment variable. Find instructions on how to create tokens and setup an environment variable in the tokens vignette (in r, send ?tokens to console).

Value

A tibble of tweets data.

See also

https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/get-statuses-lookup

Other tweets: get_favorites, get_mentions, get_timeline, lists_statuses, search_tweets, tweets_data, tweets_with_users

Examples

# NOT RUN {
## create object containing status IDs
statuses <- c(
  "567053242429734913",
  "266031293945503744",
  "440322224407314432"
)

## lookup tweets data for given statuses
tw <- lookup_statuses(statuses)
tw

## view users data for these statuses via users_data()
users_data(tw)

# }