Posts status update to user's Twitter account

Usage

post_tweet(
  status = "my first rtweet #rstats",
  media = NULL,
  token = NULL,
  in_reply_to_status_id = NULL,
  destroy_id = NULL,
  retweet_id = NULL,
  auto_populate_reply_metadata = FALSE,
  media_alt_text = NULL,
  lat = NULL,
  long = NULL,
  display_coordinates = FALSE
)

Arguments

status

Character, tweet status. Must be 280 characters or less.

media

Length 1 character vector with a file path to video media OR up-to length 4 character vector with file paths to static images to be included in tweet. The caller is responsible for managing this.

token

Use this to override authentication for a single API call. In many cases you are better off changing the default for all calls. See auth_as() for details.

in_reply_to_status_id

Status ID of tweet to which you'd like to reply. Note: in line with the Twitter API, this parameter is ignored unless the author of the tweet this parameter references is mentioned within the status text.

destroy_id

To delete a status, supply the single status ID here. If a character string is supplied, overriding the default (NULL), then a destroy request is made (and the status text and media attachments) are irrelevant.

retweet_id

To retweet a status, supply the single status ID here. If a character string is supplied, overriding the default (NULL), then a retweet request is made (and the status text and media attachments) are irrelevant.

auto_populate_reply_metadata

If set to TRUE and used with in_reply_to_status_id, leading @mentions will be looked up from the original Tweet, and added to the new Tweet from there. Defaults to FALSE.

media_alt_text

attach additional alt text metadata to the media you are uploading. Should be same length as media (i.e. as many alt text entries as there are media entries). See the official API documentation for more information.

lat

A numeric value representing the latitude of the location the tweet refers to. Range should be between -90 and 90 (north). Note that you should enable the "Precise location" option in your account via Settings and privacy > Privacy and Safety > Location. See the official Help Center section.

long

A numeric value representing the longitude of the location the tweet refers to. Range should be between -180 and 180 (west). See lat parameter.

display_coordinates

Put a pin on the exact coordinates a tweet has been sent from. Value should be TRUE or FALSE. This parameter would apply only if you have provided a valid lat/long pair of valid values.

References

Tweet: https://developer.twitter.com/en/docs/twitter-api/v1/tweets/post-and-engage/api-reference/post-statuses-update Retweet: https://developer.twitter.com/en/docs/twitter-api/v1/tweets/post-and-engage/api-reference/post-statuses-retweet-id Media: https://developer.twitter.com/en/docs/twitter-api/v1/media/upload-media/api-reference/post-media-metadata-create Alt-text: https://developer.twitter.com/en/docs/twitter-api/v1/media/upload-media/api-reference/post-media-metadata-create

See also