Twitter stop words in multiple languages data.

This data comes form a group of Twitter searches conducted at several times during the calendar year of 2017. The data are commonly observed words associated with 10 different languages, including c("ar", "en", "es", "fr", "in", "ja", "pt", "ru", "tr", "und"). Variables include "word" (potential stop words), "lang" (two or three word code), and "p" (probability value associated with frequency position along a normal distribution with higher values meaning the word occurs more frequently and lower values meaning the words occur less frequently).

stopwordslangs

Format

A tibble with three variables and 24,000 observations

Examples

head(stopwordslangs)

#> # A tibble: 6 x 3
#>   lang  word      p
#>   <chr> <chr> <dbl>
#> 1 und   a     1.000
#> 2 und   de    1.000
#> 3 und   i     1.000
#> 4 und   no    1.000
#> 5 und   3     1.000
#> 6 und   lol   1.000