Capture an image of a tweet/thread

Provide a status id or a full Twitter link to a tweet and this function will capture an image of the tweet --- or tweet + thread (if there are Twitter-linked replies) --- from the mobile version of said tweet/thread.

tweet_shot(statusid_or_url, zoom = 3, scale = TRUE)

Arguments

statusid_or_url
a valid Twitter status id (e.g. "947082036019388416") or a valid Twitter status URL (e.g. "https://twitter.com/jhollist/status/947082036019388416").

zoom
a positive number >= 1. See the help for [webshot::webshot()] for more information.

scale
auto-scale the image back to 1:1? Default it TRUE, which means magick will be used to return a "normal" sized tweet. Set it to FALSE to perform your own image manipulation.

Value

magick object

Details

For this to work, you will need to ensure the packages in Suggests: are installed as they will be loaded upon the first invocation of this function.

Use the zoom factor to get more pixels which may improve the text rendering of the tweet/thread.

Examples

if (FALSE) {
tweet_shot("947082036019388416")
tweet_shot("https://twitter.com/jhollist/status/947082036019388416")
}